pyproject.toml
[project]
name = "bioscripts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.84",
    "pandas>=2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
    "pyright>=1.1.394",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.10"
venvPath = "."
venv = ".venv"
